// src/types/grand.types.ts
export interface GrandResponse {
  name_ru: string;
  name_uz: string;
  name_en: string;
  description_uz: string;
  description_en: string;
  description_ru: string;
  from_date: string,
  to_date: string,
  id: number;
  image_url: [];
  created_at: string;
  updated_at: string;
  form_link: string;
  count_applications: number;
}